refactor: simplify text child rendering
Remove responsibility for text rendering from webstudio component
to instances tree renderer. Will simplify future collection management.
