+++ packages/react-sdk/src/index.ts
@@ -14,12 +14,7 @@ export {
   defaultStates,
 } from "./components/component-meta";
 export * from "./embed-template";
-export {
-  normalizeProps,
-  getPropsByInstanceId,
-  getInstanceIdFromComponentProps,
-  getIndexWithinAncestorFromComponentProps,
-} from "./props";
+export * from "./props";
 export { type Params, ReactSdkContext } from "./context";
 export {
   validateExpression,
