+++ packages/react-sdk/src/tree/create-elements-tree.tsx
@@ -7,10 +7,16 @@ import {
 import type { Instance, Instances } from "@webstudio-is/sdk";
 import type { Components } from "../components/components-utils";
 import { type Params, ReactSdkContext } from "../context";
-import type { WebstudioComponentProps } from "./webstudio-component";
 import type { ImageLoader } from "@webstudio-is/image";
 
 type InstanceSelector = Instance["id"][];
+
+export type WebstudioComponentProps = {
+  instance: Instance;
+  instanceSelector: Instance["id"][];
+  children: ReactNode;
+  components: Components;
+};
 
 export const createElementsTree = ({
   renderer,
@@ -73,6 +79,16 @@ export const createElementsTree = ({
   );
 };
 
+const renderText = (text: string): Array<JSX.Element> => {
+  const lines = text.split("\n");
+  return lines.map((line, index) => (
+    <Fragment key={index}>
+      {line}
+      {index < lines.length - 1 ? <br /> : null}
+    </Fragment>
+  ));
+};
+
 const createInstanceChildrenElements = ({
   instances,
   instanceSelector,
@@ -88,10 +104,10 @@ const createInstanceChildrenElements = ({
   >;
   components: Components;
 }) => {
-  const elements = [];
+  const elements: ReactNode[] = [];
   for (const child of children) {
     if (child.type === "text") {
-      elements.push(child.value);
+      elements.push(renderText(child.value));
       continue;
     }
     const childInstance = instances.get(child.value);
@@ -130,7 +146,7 @@ const createInstanceElement = ({
   Component: ForwardRefExoticComponent<
     WebstudioComponentProps & RefAttributes<HTMLElement>
   >;
-  children?: Array<JSX.Element | string>;
+  children?: ReactNode;
   components: Components;
 }) => {
   return (
