+++ packages/react-sdk/src/tree/index.ts
@@ -1,2 +1,1 @@
 export * from "./create-elements-tree";
-export * from "./webstudio-component";
